# docs/full/internal_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for flight data ingest, job status, route performance, and carrier performance. Protected routes require an API key (`Authorization: Bearer \u003ckey\u003e` or `X-API-Key`). `/health` and `/ready` are unauthenticated. Rate limits are shared in Postgres across API replicas; 429 responses include `Retry-After`, `X-RateLimit-Limit`, `X-RateLimit-Remaining`, and `X-RateLimit-Reset`.",
        "title": "flight-tracker API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/ingest/weather": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Queues one import job per month in the requested range. Provide stations explicitly, or omit stations to resolve from airport_weather_stations (ingest weather-stations first). Omit end_year/end_month for a single month. Set force=true to re-import months that already have data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ingest",
                    "internal"
                ],
                "summary": "Queue weather observation data ingest",
                "parameters": [
                    {
                        "description": "Ingest range and optional stations",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.WeatherIngestRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.WeatherIngestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.WeatherIngestConflictResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.WeatherIngestConflictResponse": {
            "type": "object",
            "properties": {
                "active_ingest_months": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.YearMonth"
                    }
                },
                "error": {
                    "type": "string"
                },
                "existing_data_months": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.YearMonth"
                    }
                }
            }
        },
        "handlers.WeatherIngestJobResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "month": {
                    "type": "integer"
                },
                "stations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "$ref": "#/definitions/model.JobStatus"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "handlers.WeatherIngestResponse": {
            "type": "object",
            "properties": {
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.WeatherIngestJobResponse"
                    }
                },
                "months_requested": {
                    "type": "integer"
                },
                "unresolved_airports": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.JobStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "JobStatusPending",
                "JobStatusRunning",
                "JobStatusCompleted",
                "JobStatusFailed"
            ]
        },
...
        "model.WeatherIngestRequest": {
            "type": "object",
            "required": [
                "start_month",
                "start_year",
                "stations"
            ],
            "properties": {
                "end_month": {
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1
                },
                "end_year": {
                    "type": "integer",
                    "minimum": 2018
                },
                "force": {
                    "type": "boolean"
                },
                "start_month": {
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1
                },
                "start_year": {
                    "type": "integer",
                    "minimum": 2018
                },
                "stations": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.YearMonth": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key as `Bearer \u003ckey\u003e`. `X-API-Key` is also accepted.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}